Pass 2 Item 5: refactor daily view into focused ritual card components
